added checker for textures and colors
